refactor: modularize subscription UI into components and integrate into settings pages
